Create README.md
This investigation is based on a CyberDefenders network forensics challenge in which a malicious macro embedded within a document initiated the download and execution of malware on a victim machine. Shortly after execution, the compromised host began communicating with an external server and ultimately exfiltrated sensitive data.

Following the detection of the suspicious activity, defenders captured the associated network traffic in a PCAP file, which serves as the primary evidence for this investigation. The objective is to thoroughly analyze the network capture to reconstruct the attacker’s actions, identify the malware’s communication behavior, and uncover indicators associated with the data exfiltration process.